+++ pkg/hardware-type-library/library.go
@@ -7,6 +7,8 @@ import (
 	"fmt"
 	"io"
 	"path"
+	"regexp"
+	"strconv"
 
 	"gopkg.in/yaml.v3"
 )
@@ -118,3 +120,79 @@ func (l *Library) GetDeviceTypesByHardwareType(hardwareType HardwareType) []Devi
 // func GetDeviceTypeBuildOut(name string) []DeviceBay {
 
 // }
+
+// TODO needs a different name
+type HardwareBuildOut struct {
+	DeviceTypeString string
+	DeviceType       DeviceType
+	Path             []string
+	Ordinal          int
+	OrdinalPath      []int
+	HardwareTypePath []HardwareType
+}
+
+// TODO make this should work the inventory data structure
+func (l *Library) GetDefaultChildHardwareBuildOut(deviceTypeString string) (results []HardwareBuildOut, err error) {
+	queue := []HardwareBuildOut{
+		{
+			DeviceTypeString: deviceTypeString,
+			Path:             []string{}, // This is the root of the path
+			Ordinal:          -1,
+		},
+	}
+
+	for len(queue) != 0 {
+		current := queue[0]
+		queue = queue[1:]
+
+		fmt.Println("Visiting: ", current.DeviceTypeString)
+		currentDeviceType, ok := l.DeviceTypes[current.DeviceTypeString]
+		if !ok {
+			panic(fmt.Sprint("Device type does not exist", current.DeviceType))
+		}
+
+		// Retrieve the hardware type at this point in time, so we only lookup in the map once
+		current.DeviceType = currentDeviceType
+		current.HardwareTypePath = append(current.HardwareTypePath, current.DeviceType.HardwareType)
+
+		for _, deviceBay := range currentDeviceType.DeviceBays {
+			fmt.Println("  Device bay:", deviceBay.Name)
+			if deviceBay.Default != nil {
+				fmt.Println("    Default:", deviceBay.Default.Slug)
+
+				// Extract the ordinal
+				// This is one way of going about, but it assumes that each name has a number
+				// There are two other ways to consider:
+				// - Embed an actual ordinal number in the yaml files
+				// - Get all of the device base with that type, and then sort them lexicographically. This is how HSM does it, but assumes the names can be sorted in a predictable order
+				r := regexp.MustCompile(`\d+`)
+				match := r.FindString(deviceBay.Name)
+				fmt.Printf("%s|%s\n", deviceBay.Name, match)
+
+				var ordinal int
+				if match != "" {
+					ordinal, err = strconv.Atoi(match)
+					if err != nil {
+						return nil, errors.Join(
+							fmt.Errorf("unable extract ordinal from device bay name (%s) from device type (%s)", deviceBay.Name, current.DeviceTypeString),
+							err,
+						)
+					}
+				}
+
+				queue = append(queue, HardwareBuildOut{
+					// Hardware type is deferred until when it is processed
+					DeviceTypeString: deviceBay.Default.Slug,
+					Path:             append(current.Path, deviceBay.Name),
+					Ordinal:          ordinal,
+					OrdinalPath:      append(current.OrdinalPath, ordinal),
+					HardwareTypePath: current.HardwareTypePath,
+				})
+			}
+		}
+
+		results = append(results, current)
+	}
+
+	return results[1:], nil
+}
